Move non-ISR UART methods to header
UART methods were split between HAL_UART.h and HAL_UART.cpp. This was necessary because global variables were used to share data with ISR. This commit changes the global variables to static public data members, allowing the non-ISR methods to be put back in the header. The ISRs remain in the CPP file.