Fix encoder halting program
Pushing encoder caused program to halt execution. Discovered this was due to the button being tied to PB0; when PB0 is driven low when it is an input, then SPI enters slave mode, and then the system halts because the slave SPI is trying to transmit, which never succeeds. Rewired button. This commit updates Schematic.h for new pinout, and makes PB0 an output.