Remove use of states
Previously, load regulator and temperature regulator passed and received info from UI module using states, which are structs containing information about the module. Since the debugger does not use this, the system behaved differently when using the debugger and when using the UI module. For consistency, and for best OOP practice, change code to use references to module instead of using states.